ch03 Spring Data JPA from h2 to postgres

+++ src/main/java/tacos/repository/OrderRepository.java
@@ -3,6 +3,6 @@ package tacos.repository;
 import org.springframework.data.repository.CrudRepository;
 import tacos.model.TacoOrder;
 
-public interface OrderRepository extends CrudRepository<TacoOrder, Long> {
+public interface OrderRepository extends CrudRepository<TacoOrder, Integer> {
 
 }
